# rustdoc-search: tighter encoding for f index